changed types

+++ client/src/main.c
@@ -1,12 +1,17 @@
 #include <arpa/inet.h>
+#include <errno.h>
+#include <getopt.h>
+#include <libgen.h>
+#include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <sys/types.h>
 #include <unistd.h>
 
 #define GROUP_NUMBER 6
 
-int PORT = 58000 + GROUP_NUMBER;
+unsigned short int PORT = 58000 + GROUP_NUMBER;
 unsigned long IP;
 
 void parseArgs(int argc, char *argv[]);
